Macros: Move from display builder into framework to allow re-use
