+++ applications/display/editor/src/main/java/org/csstudio/display/builder/editor/properties/MacrosPropertyBinding.java
@@ -13,10 +13,10 @@ import java.util.Optional;
 import org.csstudio.display.builder.editor.undo.SetWidgetPropertyAction;
 import org.csstudio.display.builder.model.Widget;
 import org.csstudio.display.builder.model.WidgetPropertyListener;
-import org.csstudio.display.builder.model.macros.Macros;
 import org.csstudio.display.builder.model.properties.MacrosWidgetProperty;
 import org.csstudio.display.builder.representation.javafx.MacrosDialog;
 import org.csstudio.display.builder.representation.javafx.ModalityHack;
+import org.phoebus.framework.macros.Macros;
 import org.phoebus.ui.dialog.DialogHelper;
 import org.phoebus.ui.undo.UndoableActionManager;
