+++ applications/display/model/src/main/java/org/csstudio/display/builder/model/macros/MacroXMLUtil.java
@@ -17,6 +17,7 @@ import javax.xml.stream.XMLOutputFactory;
 import javax.xml.stream.XMLStreamWriter;
 
 import org.csstudio.display.builder.model.persist.XMLTags;
+import org.phoebus.framework.macros.Macros;
 import org.phoebus.framework.persistence.XMLUtil;
 import org.w3c.dom.Element;
