+++ applications/display/model/src/main/java/org/csstudio/display/builder/model/properties/CommonWidgetProperties.java
@@ -16,9 +16,9 @@ import org.csstudio.display.builder.model.Widget;
 import org.csstudio.display.builder.model.WidgetProperty;
 import org.csstudio.display.builder.model.WidgetPropertyCategory;
 import org.csstudio.display.builder.model.WidgetPropertyDescriptor;
-import org.csstudio.display.builder.model.macros.Macros;
 import org.csstudio.display.builder.model.rules.RuleInfo;
 import org.csstudio.display.builder.model.widgets.ActionButtonWidget;
+import org.phoebus.framework.macros.Macros;
 import org.phoebus.vtype.VType;
 
 /** Common widget properties.
