+++ applications/display/model/src/main/java/org/csstudio/display/builder/model/widgets/ArrayWidget.java
@@ -23,11 +23,11 @@ import org.csstudio.display.builder.model.Widget;
 import org.csstudio.display.builder.model.WidgetCategory;
 import org.csstudio.display.builder.model.WidgetDescriptor;
 import org.csstudio.display.builder.model.WidgetProperty;
-import org.csstudio.display.builder.model.macros.Macros;
 import org.csstudio.display.builder.model.persist.ModelWriter;
 import org.csstudio.display.builder.model.persist.NamedWidgetColors;
 import org.csstudio.display.builder.model.persist.WidgetColorService;
 import org.csstudio.display.builder.model.properties.WidgetColor;
+import org.phoebus.framework.macros.Macros;
 
 /**
  * An Array Widget contains copies of a child widget. Each copy is assigned the
