+++ applications/display/representation-javafx/src/main/java/org/csstudio/display/builder/representation/javafx/MacrosTable.java
@@ -10,7 +10,7 @@ package org.csstudio.display.builder.representation.javafx;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 
-import org.csstudio.display.builder.model.macros.Macros;
+import org.phoebus.framework.macros.Macros;
 
 import javafx.beans.InvalidationListener;
 import javafx.beans.property.ReadOnlyStringWrapper;
