+++ applications/display/representation-javafx/src/main/java/org/csstudio/display/builder/representation/javafx/widgets/plots/ImageRepresentation.java
@@ -18,7 +18,6 @@ import org.csstudio.display.builder.model.DirtyFlag;
 import org.csstudio.display.builder.model.DisplayModel;
 import org.csstudio.display.builder.model.WidgetProperty;
 import org.csstudio.display.builder.model.WidgetPropertyListener;
-import org.csstudio.display.builder.model.macros.MacroHandler;
 import org.csstudio.display.builder.model.properties.ColorMap;
 import org.csstudio.display.builder.model.properties.PredefinedColorMaps;
 import org.csstudio.display.builder.model.util.ModelResourceUtil;
@@ -36,6 +35,7 @@ import org.csstudio.javafx.rtplot.NamedColorMappings;
 import org.csstudio.javafx.rtplot.RTImagePlot;
 import org.csstudio.javafx.rtplot.RTImagePlotListener;
 import org.csstudio.javafx.rtplot.RegionOfInterest;
+import org.phoebus.framework.macros.MacroHandler;
 import org.phoebus.util.array.ArrayDouble;
 import org.phoebus.util.array.ArrayInt;
 import org.phoebus.vtype.VImage;
