+++ core/framework/src/main/java/org/phoebus/framework/macros/MacroHandler.java
@@ -1,11 +1,11 @@
 /*******************************************************************************
- * Copyright (c) 2015-2016 Oak Ridge National Laboratory.
+ * Copyright (c) 2015-2018 Oak Ridge National Laboratory.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *******************************************************************************/
-package org.csstudio.display.builder.model.macros;
+package org.phoebus.framework.macros;
 
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.regex.Matcher;
